+++ SFRecordCompareEngine/Services/Interfaces/IApplicationNavigationService.cs
@@ -3,6 +3,7 @@ namespace SFRecordCompareEngine.Services.Interfaces;
 public interface IApplicationNavigationService
 {
     Task ShowMainPageAsync();
+    Task ShowStartupImportAsync(bool forceFullReimport);
     Task ShowOpenDialogAsync();
     Task ShowSettingsDialogAsync();
     Task CloseOpenDialogAsync();